GEN-24070 Species delimitation crashes when tree re-rooted (group 10078)
Make it so the species delimitation panel never updates when the tree is edited in the viewer, you must now click the Reload button or use the controls in the panel to add groups. This was really the intention all along but the listeners were misbehaving.

Specifically I removed the restart() call when a crash is encountered in the selection listener which always happens when a tree is edited.